Refactor mistdemo web: WebDemo→Web rename, response split, lint cleanup
- Rename WebDemo* types and files to Web* (WebDemoServer→WebServer,
  WebDemoBackend→WebBackend, WebDemoIndexHTML→WebIndexHTML, etc.) and
  drop the webDemo* prefix from WebBackend's protocol methods.
- Split WebDemoRequests into WebRequests (Query/Create/Update/Delete)
  and a new WebResponse (Records/Delete/Error); drop the redundant
  *Request/*Response suffixes from the inner type names.
- Extract WebBackendFactory into its own file (one_declaration_per_file).
- Move WebServer's CRUD endpoint adders into WebServer+CRUD.swift and
  reorder type contents so static helpers precede instance methods
  (file_length + type_contents_order under --strict).
- Split WebServerTests CRUD tests into WebServerTests+CRUD.swift to fit
  file_length / type_body_length; drop unused encode-placeholders that
  triggered multiline_arguments_brackets.
- AuthTokenCommand: extract captureToken static helper for
  function_body_length; reorder to subtype → init → type_method →
  instance_method.
- MockBackend: add explicit internal ACLs (13) and reorder so type
  methods precede instance methods.
- WebAuthTokenStore: reorder properties → init → methods → deinit.
- WebIndexHTML: replace try!/force-unwrap with precondition in loader.
- TestIntegrationCommand / TestPrivateCommand: reformat --lookup-email
  help text to use 2-space indent on continuation lines.
- MistDemoConfig+DatabaseConfiguration: move computed property before
  instance method.
- WebCommand: split guard return onto own line.

Branch now passes ./Scripts/lint.sh under LINT_MODE=STRICT (45 → 0
swiftlint violations); 918 tests pass.
