+++ Examples/MistDemo/Sources/MistDemoKit/Commands/TestIntegrationCommand.swift
@@ -55,10 +55,10 @@ public struct TestIntegrationCommand: MistDemoCommand {
       --asset-size <kb>        Asset size in KB (default: 100)
       --skip-cleanup           Skip cleanup after test
       --verbose                Run in verbose mode
-      --lookup-email <email>   Email for users/lookup/email phase
-                               (CLOUDKIT_LOOKUP_EMAIL); must belong
-                               to an iCloud account discoverable to
-                               the caller, otherwise the phase skips
+      --lookup-email <email>
+        Email for users/lookup/email phase (CLOUDKIT_LOOKUP_EMAIL).
+        Must belong to an iCloud account discoverable to the caller,
+        otherwise the phase skips.
 
     EXAMPLES:
       mistdemo test-integration --verbose
