+++ Examples/MistDemo/Sources/MistDemoKit/Server/WebIndexHTML.swift
@@ -1,5 +1,5 @@
 //
-//  WebDemoIndexHTML.swift
+//  WebIndexHTML.swift
 //  MistDemo
 //
 //  Created by Leo Dion.
@@ -30,19 +30,28 @@
 #if canImport(Hummingbird)
   internal import Foundation
 
-  /// Loader for the web-demo's interactive page served by `WebDemoServer`.
+  /// Loader for the web command's interactive page served by `WebServer`.
   ///
   /// The HTML+JS lives in `Resources/index.html` and is read from
   /// `Bundle.module` on first access. The mode toggle in this page lets
   /// users compare MistKit (server-side) and CloudKit JS (browser-side)
   /// against the same CloudKit container; the CloudKit JS side is wired
   /// in by #329.
-  internal enum WebDemoIndexHTML {
-    internal static let content: String = try! String(
-      contentsOf: Bundle.module.url(
-        forResource: "index", withExtension: "html"
-      )!,
-      encoding: .utf8
-    )
+  internal enum WebIndexHTML {
+    internal static let content: String = loadContent()
+
+    private static func loadContent() -> String {
+      guard
+        let url = Bundle.module.url(
+          forResource: "index", withExtension: "html"
+        ),
+        let html = try? String(contentsOf: url, encoding: .utf8)
+      else {
+        preconditionFailure(
+          "Resources/index.html missing from MistDemoKit bundle"
+        )
+      }
+      return html
+    }
   }
 #endif
