+++ Examples/MistDemo/Sources/MistDemoKit/Server/WebServer+CRUD.swift
@@ -1,0 +1,138 @@
+//
+//  WebServer+CRUD.swift
+//  MistDemo
+//
+//  Created by Leo Dion.
+//  Copyright © 2026 BrightDigit.
+//
+//  Permission is hereby granted, free of charge, to any person
+//  obtaining a copy of this software and associated documentation
+//  files (the "Software"), to deal in the Software without
+//  restriction, including without limitation the rights to use,
+//  copy, modify, merge, publish, distribute, sublicense, and/or
+//  sell copies of the Software, and to permit persons to whom the
+//  Software is furnished to do so, subject to the following
+//  conditions:
+//
+//  The above copyright notice and this permission notice shall be
+//  included in all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+//  OTHER DEALINGS IN THE SOFTWARE.
+//
+
+#if canImport(Hummingbird)
+  internal import Foundation
+  internal import Hummingbird
+  internal import MistKit
+
+  extension WebServer {
+    internal func addQueryEndpoint(
+      api: RouterGroup<BasicRequestContext>
+    ) {
+      let tokenStore = self.tokenStore
+      let backendFactory = self.backendFactory
+      api.post("records/query") { request, context -> Response in
+        guard let token = await tokenStore.currentToken else {
+          return Response(status: .unauthorized)
+        }
+        let body = try await request.decode(
+          as: WebRequests.Query.self, context: context
+        )
+        return try await Self.runOperation { () -> Data in
+          let backend = try backendFactory.make(token)
+          let records = try await backend.webQuery(
+            recordType: body.recordType, limit: body.limit
+          )
+          return try JSONEncoder().encode(
+            WebResponse.Records(records: records)
+          )
+        }
+      }
+    }
+
+    internal func addCreateEndpoint(
+      api: RouterGroup<BasicRequestContext>
+    ) {
+      let tokenStore = self.tokenStore
+      let backendFactory = self.backendFactory
+      api.post("records/create") { request, context -> Response in
+        guard let token = await tokenStore.currentToken else {
+          return Response(status: .unauthorized)
+        }
+        let body = try await request.decode(
+          as: WebRequests.Create.self, context: context
+        )
+        return try await Self.runOperation { () -> Data in
+          let backend = try backendFactory.make(token)
+          let record = try await backend.webCreate(
+            recordType: body.recordType,
+            fields: WebRequests.stringFields(body.fields)
+          )
+          return try JSONEncoder().encode(
+            WebResponse.Records(records: [record])
+          )
+        }
+      }
+    }
+
+    internal func addUpdateEndpoint(
+      api: RouterGroup<BasicRequestContext>
+    ) {
+      let tokenStore = self.tokenStore
+      let backendFactory = self.backendFactory
+      api.post("records/update") { request, context -> Response in
+        guard let token = await tokenStore.currentToken else {
+          return Response(status: .unauthorized)
+        }
+        let body = try await request.decode(
+          as: WebRequests.Update.self, context: context
+        )
+        return try await Self.runOperation { () -> Data in
+          let backend = try backendFactory.make(token)
+          let record = try await backend.webUpdate(
+            recordType: body.recordType,
+            recordName: body.recordName,
+            fields: WebRequests.stringFields(body.fields)
+          )
+          return try JSONEncoder().encode(
+            WebResponse.Records(records: [record])
+          )
+        }
+      }
+    }
+
+    internal func addDeleteEndpoint(
+      api: RouterGroup<BasicRequestContext>
+    ) {
+      let tokenStore = self.tokenStore
+      let backendFactory = self.backendFactory
+      api.post("records/delete") { request, context -> Response in
+        guard let token = await tokenStore.currentToken else {
+          return Response(status: .unauthorized)
+        }
+        let body = try await request.decode(
+          as: WebRequests.Delete.self, context: context
+        )
+        return try await Self.runOperation { () -> Data in
+          let backend = try backendFactory.make(token)
+          try await backend.webDelete(
+            recordType: body.recordType,
+            recordName: body.recordName
+          )
+          return try JSONEncoder().encode(
+            WebResponse.Delete(
+              recordName: body.recordName, deleted: true
+            )
+          )
+        }
+      }
+    }
+  }
+#endif
